Created DAO, rendered POST request to create a new account

+++ Main/src/main/java/com/project/Main/repositories/ClientRepository.java
@@ -3,6 +3,5 @@ package com.project.Main.repositories;
 import com.project.Main.models.Client;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClientRepository extends JpaRepository<Client, Integer> {
-
+public interface ClientRepository extends JpaRepository<Client, Long> {
 }
